% Generated by roxygen2: do not edit by hand
\docType{class}
\name{opt_fwhm}
\title{ARI Spatial smoothing - Function optimizing the smoothing parameter within the All-Resolutions Inference framework}
\value{
Returns a value for the optimal smoothing parameter within the All-Resolutions Inference framework.
}
\description{
Function that calculates based on the optimization choice either the maximum detected cluster size or
number of clusters that are found given a TDP threshold when applying the All-Resolutions Inference framework.
}


\author{
Xu Chen, Thijmen Krebs, Wouter Weeda.
}
