% Generated by roxygen2: do not edit by hand
\name{optfwhm}
\title{ARI Spatial smoothing - Function optimizing the smoothing parameter within the All-Resolutions Inference framework}
\usage{
optfwhm(
  zstat,
  min_fwhm,
  max_fwhm,
  tdp,
  max_size
)
}
\arguments{
\item{zstat}{Z-stats 3D array of activation values, or a nifti file name.}

\item{min_fwhm}{Minimum level of spatial smoothing a researcher wishes to set, default is 2.}

\item{max_fwhm}{Maximum level of spatial smoothing a researcher wishes to set, default is 2.}

\item{tdp}{Minimum TDP threshold set}

\item{max_size}{Optimization based on maximal size (True) or maximal number of clusters (False)}
}
\value{
Returns a value that can be used to define the optimal smoothing parameter.
}
\description{
Function that calculates based on the optimization choice either the maximum detected cluster size or
number of clusters that are found given a TDP threshold when applying the All-Resolutions Inference framework.
}

\examples{

optfwhm(zstat, min_fwhm = 2, max_fwhm = 10, tdp = 0.7, max_size = T)
    
}
